backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai>=1.30.0
httpx>=0.28.1
python-dotenv==1.0.0
pydantic>=2.6.0
python-multipart==0.0.6
